Moved Windows-only Permission function to Radarr.Windows
